requirements/base.txt
Django==3.0.7
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
django-extensions==3.1.0

whitenoise==5.1.0

django-filter==2.4.0
drf-yasg
